feat(auth-refactor): parse magic link to utils
